requirements.txt
# Core dependencies
asyncio-mqtt>=0.13.0
aiofiles>=23.2.1
aiohttp>=3.9.0
aiosqlite>=0.19.0
PyYAML>=6.0.1
prometheus-client>=0.19.0
web3>=6.11.0

# Additional utilities
python-dateutil>=2.8.0
click>=8.1.0

# Monitoring and metrics
psutil>=5.9.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
mypy>=1.6.0
